test: rename contexts and simplify.
