+++ test/node-rescan-test.js
@@ -6,7 +6,7 @@ const {BloomFilter} = require('@handshake-org/bfilter');
 const TX = require('../lib/primitives/tx');
 const nodeCommon = require('../lib/blockchain/common');
 const {scanActions} = nodeCommon;
-const NodeContext = require('./util/node');
+const NodeContext = require('./util/node-context');
 const {forEvent, sleep} = require('./util/common');
 const MemWallet = require('./util/memwallet');
 const rules = require('../lib/covenants/rules');
